refactor(ycsb_thread): read stats from log
Since i've found a way to limit ycsb by time, no more need
for running it in small batches in a loop

Also refactor and there a new base class for the docker based stress tools
