+++ data_dir/scylla-dash-per-server-nemesis.2019.1.json
@@ -773,21 +773,21 @@
                                 "refId": "F"
                             },
                             {
-                                "expr": "collectd_ycsb_read_gauge{type=\"lat_perc_95\"}",
+                                "expr": "collectd_ycsb_read_gauge{type=\"p90\"}",
                                 "format": "time_series",
                                 "interval": "15s",
                                 "intervalFactor": 1,
                                 "refId": "G"
                             },
                             {
-                                "expr": "collectd_ycsb_update_gauge{type=\"lat_perc_95\"}",
+                                "expr": "collectd_ycsb_update_gauge{type=\"p90\"}",
                                 "format": "time_series",
                                 "interval": "15s",
                                 "intervalFactor": 1,
                                 "refId": "H"
                             },
                                                         {
-                                "expr": "collectd_ycsb_cleanup_gauge{type=\"lat_perc_95\"}",
+                                "expr": "collectd_ycsb_insert_gauge{type=\"p90\"}",
                                 "format": "time_series",
                                 "interval": "15s",
                                 "intervalFactor": 1,
@@ -915,21 +915,21 @@
                                 "refId": "F"
                             },
                             {
-                                "expr": "collectd_ycsb_read_gauge{type=\"lat_perc_99\"}",
+                                "expr": "collectd_ycsb_read_gauge{type=\"p99\"}",
                                 "format": "time_series",
                                 "interval": "15s",
                                 "intervalFactor": 1,
                                 "refId": "G"
                             },
                             {
-                                "expr": "collectd_ycsb_update_gauge{type=\"lat_perc_99\"}",
+                                "expr": "collectd_ycsb_update_gauge{type=\"p99\"}",
                                 "format": "time_series",
                                 "interval": "15s",
                                 "intervalFactor": 1,
                                 "refId": "H"
                             },
                                                         {
-                                "expr": "collectd_ycsb_cleanup_gauge{type=\"lat_perc_99\"}",
+                                "expr": "collectd_ycsb_insert_gauge{type=\"p99\"}",
                                 "format": "time_series",
                                 "interval": "15s",
                                 "intervalFactor": 1,
