+++ sdcm/ycsb_thread.py
@@ -1,36 +1,30 @@
 import os
+import re
 import logging
 import time
 import uuid
-import concurrent.futures
 import tempfile
 from textwrap import dedent
 
 from sdcm.prometheus import nemesis_metrics_obj
 from sdcm.sct_events import YcsbStressEvent
 from sdcm.remote import FailuresWatcher
-from sdcm.ndbench_thread import RemoteDocker
+from sdcm.utils.common import FileFollowerThread
+from sdcm.utils.thread import DockerBasedStressThread
+from sdcm.utils.docker import RemoteDocker
 
 LOGGER = logging.getLogger(__name__)
 
 
-class YcsbStressThread():  # pylint: disable=too-many-instance-attributes
+class YcsbStatsPublisher(FileFollowerThread):
     METRICS = dict()
     collectible_ops = ['read', 'update', 'cleanup', 'read-failed', 'update-failed']
 
-    def __init__(self, loader_set, stress_cmd, timeout, stress_num=1, node_list=None, round_robin=False,  # pylint: disable=too-many-arguments
-                 params=None):
-        self.loader_set = loader_set
-        self.stress_cmd = stress_cmd
-        self.timeout = timeout
-        self.stress_num = stress_num
-        self.node_list = node_list if node_list else []
-        self.round_robin = round_robin
-        self.params = params if params else dict()
-
-        self.executor = None
-        self.results_futures = []
-        self.max_workers = 0
+    def __init__(self, loader_node, loader_idx, ycsb_log_filename):
+        super().__init__()
+        self.loader_node = loader_node
+        self.loader_idx = loader_idx
+        self.ycsb_log_filename = ycsb_log_filename
 
         for operation in self.collectible_ops:
             gauge_name = self.gauge_name(operation)
@@ -44,23 +38,51 @@ class YcsbStressThread():  # pylint: disable=too-many-instance-attributes
     def gauge_name(operation):
         return 'collectd_ycsb_%s_gauge' % operation.replace('-', '_')
 
+    def set_metric(self, operation, name, value):
+        metric = self.METRICS[self.gauge_name(operation)]
+        metric.labels(self.loader_node.ip_address, self.loader_idx, name).set(value)
+
     def run(self):
-        if self.round_robin:
-            self.stress_num = 1
-            loaders = [self.loader_set.get_loader()]
-            LOGGER.debug("Round-Robin through loaders, Selected loader is {} ".format(loaders))
-        else:
-            loaders = self.loader_set.nodes
+        # pylint: disable=too-many-nested-blocks
 
-        self.max_workers = len(loaders) * self.stress_num
-        LOGGER.debug("Starting %d ycsb Worker threads", self.max_workers)
-        self.executor = concurrent.futures.ThreadPoolExecutor(max_workers=self.max_workers)
+        # 729.39 current ops/sec;
+        # [READ: Count=510, Max=195327, Min=2011, Avg=4598.69, 90=5743, 99=12583, 99.9=194815, 99.99=195327]
+        # [CLEANUP: Count=5, Max=3, Min=0, Avg=0.6, 90=3, 99=3, 99.9=3, 99.99=3]
+        # [UPDATE: Count=490, Max=190975, Min=2004, Avg=3866.96, 90=4395, 99=6755, 99.9=190975, 99.99=190975]
 
-        for loader_idx, loader in enumerate(loaders):
-            for cpu_idx in range(self.stress_num):
-                self.results_futures += [self.executor.submit(self._run_stress, *(loader, loader_idx, cpu_idx))]
+        regex_dict = dict()
+        for operation in self.collectible_ops:
+            regex_dict[operation] = re.compile(
+                fr'\[{operation.upper()}:\sCount=(?P<count>\d*?),'
+                fr'.*?Max=(?P<max>\d*?),.*?Min=(?P<min>\d*?),'
+                fr'.*?Avg=(?P<avg>.*?),.*?90=(?P<p90>\d*?),'
+                fr'.*?99=(?P<p99>\d*?),.*?99.9=(?P<p999>\d*?),'
+                fr'.*?99.99=(?P<p9999>\d*?)\]'
+            )
 
-        return self
+        while not self.stopped():
+            exists = os.path.isfile(self.ycsb_log_filename)
+            if not exists:
+                time.sleep(0.5)
+                continue
+
+            for _, line in enumerate(self.follow_file(self.ycsb_log_filename)):
+                if self.stopped():
+                    break
+                try:
+                    for operation, regex in regex_dict.items():
+                        match = regex.search(line)
+                        if match:
+                            for key, value in match.groupdict().items():
+                                if not key == 'count':
+                                    value = float(value) / 1000.0
+                                self.set_metric(operation, key, float(value))
+
+                except Exception:  # pylint: disable=broad-except
+                    LOGGER.exception("fail to send metric")
+
+
+class YcsbStressThread(DockerBasedStressThread):  # pylint: disable=too-many-instance-attributes
 
     def copy_template(self, docker):
         if 'dynamodb' in self.stress_cmd:
@@ -101,106 +123,42 @@ class YcsbStressThread():  # pylint: disable=too-many-instance-attributes
                 tmp_file.flush()
                 docker.send_files(tmp_file.name, os.path.join('/tmp', 'aws_empty_file'))
 
-            self.stress_cmd += ' -P /tmp/dynamodb.properties'
-
-    def create_stress_cmd(self, node, loader_idx):
-        pass
+            self.stress_cmd += f' -s -P /tmp/dynamodb.properties -p maxexecutiontime={self.timeout}'
 
     def _run_stress(self, loader, loader_idx, cpu_idx):
-        # pylint: disable=too-many-locals, too-many-nested-blocks, too-many-branches
-        output = ''
-        docker = RemoteDocker(loader, "scylladb/hydra-loaders:ycsb-jdk8-20200130")
+
+        docker = RemoteDocker(loader, "scylladb/hydra-loaders:ycsb-jdk8-20200130",
+                              extra_docker_opts=f'--network=host --label shell_marker={self.shell_marker}')
         self.copy_template(docker)
 
-        if not os.path.exists(loader.logdir):
-            os.makedirs(loader.logdir)
-        log_file_name = os.path.join(loader.logdir, 'cassandra-stress-l%s-c%s-%s.log' %
+        log_dir = os.path.join(loader.logdir, self.loader_set.name)
+        if not os.path.exists(log_dir):
+            os.makedirs(log_dir)
+        log_file_name = os.path.join(log_dir, 'ycsb-l%s-c%s-%s.log' %
                                      (loader_idx, cpu_idx, uuid.uuid4()))
-        LOGGER.debug('cassandra-stress local log: %s', log_file_name)
-
-        def set_metric(operation, name, value):
-            metric = self.METRICS[self.gauge_name(operation)]
-            metric.labels(loader.ip_address, loader_idx, name).set(value)
-
-        def inc_metric(operation, name, value):
-            metric = self.METRICS[self.gauge_name(operation)]
-            metric.labels(loader.ip_address, loader_idx, name).inc(value)
+        LOGGER.debug('ycsb-stress local log: %s', log_file_name)
 
         def raise_event_callback(sentinal, line):  # pylint: disable=unused-argument
             if line:
                 YcsbStressEvent('error', node=loader, stress_cmd=self.stress_cmd, errors=[line])
 
         YcsbStressEvent('start', node=loader, stress_cmd=self.stress_cmd)
-        timeout_start = time.time()
         try:
-            while time.time() < timeout_start + self.timeout:
+            LOGGER.debug("running: %s", self.stress_cmd)
 
-                LOGGER.debug("running: %s", self.stress_cmd)
+            if self.stress_num > 1:
+                node_cmd = 'taskset -c %s bash -c "%s"' % (cpu_idx, self.stress_cmd)
+            else:
+                node_cmd = self.stress_cmd
 
-                if self.stress_num > 1:
-                    node_cmd = 'taskset -c %s bash -c "%s"' % (cpu_idx, self.stress_cmd)
-                else:
-                    node_cmd = self.stress_cmd
+            node_cmd = 'cd /YCSB && {}'.format(node_cmd)
 
-                node_cmd = 'cd /YCSB && {}'.format(node_cmd)
+            with YcsbStatsPublisher(loader, loader_idx, ycsb_log_filename=log_file_name):
                 result = docker.run(cmd=node_cmd,
-                                    timeout=self.timeout,
-                                    ignore_status=True,
+                                    timeout=self.timeout + 60,
                                     log_file=log_file_name,
                                     watchers=[FailuresWatcher('ERROR', callback=raise_event_callback)])
-
-                assert result.exit_status == 0
-
-                LOGGER.debug(result.stdout)
-                output += result.stdout
-                run_time = 0
-                # TODO: get statistic from -s format
-                #  '''2019-08-06 13:03:18:305 10 sec: 5390 operations; 538.95 current ops/sec; est completion in 1 day 1 hour [INSERT: Count=5391, Max=1956863, Min=3196, Avg=204978.3, 90=324863, 99=1863679, 99.9=1902591, 99.99=1927167] '''
-                for line in result.stdout.splitlines():
-                    if '[OVERALL]' in line:
-                        title, _type, measure = line.split(', ')
-                        if title == '[OVERALL]' and _type == 'RunTime(ms)':
-                            run_time = int(measure)
-
-                    for operation in self.collectible_ops:
-                        if '[{}]'.format(operation.upper()) in line:
-                            title, _type, measure = line.split(', ')
-                            if _type == 'Operations':
-                                set_metric(operation, 'ops', float(measure) / (run_time / 1000.0))
-                            if _type == 'Return=CLIENT_ERROR':
-                                LOGGER.error("%s %s errors were found", _type, measure)
-                                inc_metric(operation, 'errors', float(measure))
-
-                            short_name_mapping = {
-                                'MaxLatency(us)': 'lat_max',
-                                '95thPercentileLatency(us)': 'lat_perc_95',
-                                '99thPercentileLatency(us)': 'lat_perc_99',
-                                'AverageLatency(us)': 'lat_med',
-                            }
-                            if _type in short_name_mapping:
-                                set_metric(operation, short_name_mapping[_type], float(measure) * 0.001)
         finally:
             YcsbStressEvent('finish', node=loader, stress_cmd=self.stress_cmd, log_file_name=log_file_name)
 
-        return output
-
-    def get_results(self):
-        ret = []
-        results = []
-        timeout = self.timeout + 120
-        LOGGER.debug('Wait for %s stress threads results', self.max_workers)
-        for future in concurrent.futures.as_completed(self.results_futures, timeout=timeout):
-            results.append(future.result())
-
-        return ret
-
-    def verify_results(self):
-        ret = []
-        results = []
-        errors = []
-        timeout = self.timeout + 120
-        LOGGER.debug('Wait for %s stress threads to verify', self.max_workers)
-        for future in concurrent.futures.as_completed(self.results_futures, timeout=timeout):
-            results.append(future.result())
-
-        return ret, errors
+        return result
